+++ src/main/java/com/renomad/minum/web/BodyProcessor.java
@@ -279,8 +279,12 @@ final class BodyProcessor implements IBodyProcessor {
                     String s;
                     try {
                         s = inputStreamUtils.readLine(inputStream);
+                        if (s == null) {
+                            throw new IOException("Unexpectedly encountered end of stream while reading in BodyProcessor.next()");
+                        }
                         countBytesRead.incrementBy(s.length() + 2);
                         hasReadFirstPartition = true;
+
                         if (!s.contains(boundaryValue)) {
                             throw new IOException("Error: First line must contain the expected boundary value. Expected to find: "+ boundaryValue + " in: " + s);
                         }