+++ src/main/java/com/renomad/minum/web/WebFramework.java
@@ -17,7 +17,6 @@ import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.*;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
@@ -110,7 +109,7 @@ public final class WebFramework {
                 while (true) {
                     final String rawStartLine = inputStreamUtils.readLine(is);
                     long startMillis = System.currentTimeMillis();
-                    if (rawStartLine.isEmpty()) {
+                    if (rawStartLine == null || rawStartLine.isEmpty()) {
                         // here, the client connected, sent nothing, and closed.
                         // nothing to do but return.
                         logger.logTrace(() -> "rawStartLine was empty.  Returning.");
@@ -729,9 +728,18 @@ public final class WebFramework {
      * <br>
      * Note that the path text expected is *after* the first forward slash,
      * so for example with {@code http://foo.com/mypath}, provide "mypath" as the path.
+     * @throws WebServerException if duplicate paths are registered, or if the path is prefixed with a slash
      */
     public void registerPath(RequestLine.Method method, String pathName, ThrowingFunction<IRequest, IResponse> webHandler) {
-        registeredDynamicPaths.put(new MethodPath(method, pathName), webHandler);
+        if (pathName.startsWith("\\") || pathName.startsWith("/")) {
+            throw new WebServerException(
+                    String.format("Path should not be prefixed with a slash.  Corrected version: registerPath(%s, \"%s\", ... )", method.name(), pathName.substring(1)));
+        }
+
+        var result = registeredDynamicPaths.put(new MethodPath(method, pathName), webHandler);
+        if (result != null) {
+            throw new WebServerException("Duplicate endpoint registered: " + new MethodPath(method, pathName));
+        }
     }
 
     /**
@@ -745,9 +753,17 @@ public final class WebFramework {
      *     Be careful here, be thoughtful - partial paths will match a lot, and may
      *     overlap with other URL's for your app, such as endpoints and static files.
      * </p>
+     * @throws WebServerException if duplicate paths are registered, or if the path is prefixed with a slash
      */
     public void registerPartialPath(RequestLine.Method method, String pathName, ThrowingFunction<IRequest, IResponse> webHandler) {
-        registeredPartialPaths.put(new MethodPath(method, pathName), webHandler);
+        if (pathName.startsWith("\\") || pathName.startsWith("/")) {
+            throw new WebServerException(
+                    String.format("Path should not be prefixed with a slash.  Corrected version: registerPartialPath(%s, \"%s\", ... )", method.name(), pathName.substring(1)));
+        }
+        var result = registeredPartialPaths.put(new MethodPath(method, pathName), webHandler);
+        if (result != null) {
+            throw new WebServerException("Duplicate partial-path endpoint registered: " + new MethodPath(method, pathName));
+        }
     }
 
     /**